Unify job traversal for qstat -u and qstat -j
There are subtle differences both in the contents of the jobs
("fields") and in the outer structure (names of tags). Instead of the
old, ugly hack in Jobs.Append_List, we now have
Parser.Get_Job_Nodes_From_Qstat_{U,J}. This Subprogram will (almost)
generate a list of "job" nodes (actually, they are called job_list and
element, resp.) for a simplified Append_List to process. "Almost"
means that in the case of qstat -j, #text nodes are interspersed and
have to be removed by Append_List.

Oh, and the old Append_List would ignore everything but the first job
for qstat -j.