Start Resources.Hashed_List type
This type will keep a hash value of the whole list (actually a map)
and track all insertions to make sure it stays valid (or better yet,
mark as invalid and recalculate as required).
We use a tagged type with private, overriding subprograms for Insert
etc. In this way, we make sure that no-one bypasses our
hash-invalidating subprograms.