Update jobs and bunches with dual slots
Slots are reported by SGE (and tracked by qview) either as a scalar,
or as a list of ranges (possibly with step sizes). After some
consideration, I've found that the cleanest way to handle this is to
keep track of both, but allow for the possibility that either may be
empty.