Handle Job Args
Some jobs have JB_job_args. These are given (to qsub) at submission
time. At run time, they are available to the job script. Previously,
the presence of JB_job_args would trigger a warning in the CGI. Now,
args are stored with the job. They are presented in job view (in the
files pane), and ignored otherwise.