Refactor: Inappropriate Intimacy w/r to Jobs
The Jobs package is far too public. This can easily be seen when
recompiling after jobs.ads has been changed -- at least half the
project is affected.
I've moved the Job record (and, therefore, the Job_Lists package) into
the private part. The Viewer package has been changed accordingly
(although there are still several subprograms that should be moved
elsewhere in the long run); next up is the Bunches package: this is a
trifle more difficult, because Bunches need to know a lot about Jobs.
I will probably need quite a few accessors.

Rationale for Bunches.Build_List (in preference to
Jobs.Build_Bunch_List): a Bunch is a representation of one or more
Jobs. Thus it is only natural that it needs to know about Jobs.
However, a Job does not need to know anything about Bunches. It is
merely a part of a Bunch, just as an item is a part of a list: the
list needs to know a few things about the item (such as relative
ordering, or how to create and destroy items), but items are
independent of lists.