Towards reservation analysis
Qview still does nothing for reservations (-R y, not advance reservations)
because SGE tools cannot be used to analyze them. However, such knowledge
is highly desirable if not for users, at least for admins in order to adjust
the cluster policy. So, read the schedule log directly.
This commit starts implementing that. Reading is finished, output is not.
The main problem is that the schedule log grows so fast -- half a million lines
per day are not uncommon.
We could implement log rotation (but we would need to depend on an outside factor),
or else seek from the end of the file. This could even be adjustable by the
caller.