Keep track of Job_IDs to fix #1280
We no longer show "starting" entries for jobs without a reservation.
This is done via a set that records Job_IDs. That set is also used to
distinguish between new and shifted reservations. Note that multi-node
reservations are shown as new only for the first node. The others are
shown as shifted since the job has been seen, but there is no matching
reservation in the previous iteration.
This should be fixed with #1281.

Also some cosmetic changes.