Reservations: Iteration numbers added (#1284)
We now have an additional column with the iteration number. We also
hide inner confirmations, i.e. every "confirmed" reservation that is
followed by a confirmation for the same Job_ID in the next iteration
is hidden. This reduces the screen clutter quite nicely, so we can
actually follow what's going on inside the scheduler.