Refactoring: #1208
Move code from HTML to Jobs and Partitions:
In an earlier stage of development, a lot of code to present internal
structures from various packages would end up in the HTML package.
Now, the last subprograms (pertaining to states and memory) have been
moved to the packages that contain the type definitions.