Fix partitions with GPU/SSD
Previously, clicking on the "detail" arrow of a partition with SSDs
(or GPUs) in "detailed queues" view would result in an empty list.
We now generate correct links, so the host list is shown with all
nodes. This is different from #1429, where an error message appeared
when looking at these hosts.