Report hosts as well as slots (#1441)
Up to now, only slots (free, reserved, used, ...) would be reported in
supply view. The numbers are often confusing (because of their
magnitude), especially with the newer machines with 48/64 cores.
Moreover, one cannot see how the free slots are distributed across the
nodes. For a detailed view, one can always click the arrow to got to
Host view, but knowing how many hosts are (completely) free already
gives an overview.