Refactor: Viewer.View_Hosts
In preparation of #1484, pull out the parameter reading logic from
View_Hosts and create a new procedure View_Partition that initializes
a Set_Of_Properties from the cgi parameters.
The new View_Hosts is then called with this Set_Of_Properties.
When implementing #1484, we can compose the Set_Of_Properties in a
different way and still use View_Hosts.