Implement CSS (Bug #1500)
This first version reproduces the complete (almost) style sheet, but
employs little automation.
We have Open_Block and Close_Block with automatic indents, and a few
constants for colours, borders, and widths, but that is about it.
Some more structure would be nice, but is best introduced when the
style sheet actually changes.