Some trimming of job lists
In order to make more room for job names (Bug #1387), we suppress CPU
minutes and seconds whenever more than one CPU day has been used. We
suppress all resource columns if only waiting jobs are displayed.

Also, show 25 instead of 20 characters in the job name column.
On the other hand, it is the user's responsibility to generate
sensible job names (if 25 characters are not enough, who says 250
are?), or to use the job ID to identify jobs.
Priorities may not be relevant to everyone, but they are useful
to understand the order of job execution.

Is there any process that can be automated so that manual job
identification becomes irrelevant, e.g. by adding functionality to
qview?