Add "old" queue state (Bug #1602)
This state occurs when a queue has been removed from a host, but
cannot be done away with because it still has running jobs on that
host.
This happens only rarely, so it slipped by until now.
For now, we do not do anything with the information collected (unlike
disabled or error).