Finish change to Countable_Sets and _Maps
Progressing with Bug #1654, all Slots (in a partition, and in the
summary) are now counted by Countable_Maps. The only exception are
Used and Reserved slots: These are real, two slots used via queue A
and three via queue B mean five are used on that host.

Unfortunately, this exception *does not* apply to the Summary line.

Also, all Hosts are counted by Countable_Sets.