Add full support for suspended queues (Bug #1654)
Count (S)uspended slots for each partition (but not in the summary).
This is done via a scalar: suspended slots are "virtual" anyway, and
it does not make any sense to report 8 suspended slots if we have 3
queues on an 8-core node when that node is fully loaded via one of the
queues: no "real" slot is affected (i.e. nothing is idle), so be
honest and report 16 slots.
For the same reason, it does not make sense to count "suspended
hosts": only queues can be suspended, and this only happens if jobs
run on other queues *on the same host*.