Fix Bug #1637
For completed jobs, "usage" contans very many fields that have been
unsupported until now. This is mainly relevant for array jobs: they
are around for a long time even after some tasks are completed.
We add a few fields to the appropriate type and ignore the others
based on string comparison.